Task 0

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -8,7 +8,7 @@
 
 hash_table_t *hash_table_create(unsigned long int size)
 {
-	unsigned long int i;
+	unsigned int i;
 	hash_table_t *h_table = malloc(sizeof(hash_table_t));
 
 	if (h_table == NULL)
